requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-filter==23.2
django-import-export==3.2.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.25.1
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
